Refactor: Split SupportSettingsManager component
Split the SupportSettingsManager component into smaller, more manageable components for better maintainability.
